Navigation to a new page will scroll to top. Fixes #7
