+++ app/src/js/entry/router/index.js
@@ -1,0 +1,3 @@
+import router from './router';
+
+export default router;
